refactor: reorganize utilities and broker export
